Move collectAncestorDirectories into CLI lib
collectAncestorDirectories previously lived in src/shared/path-utils.js but only the CLI relies on it for .prettierignore discovery. Move it into src/cli/lib/ancestor-directories.js so shared remains focused on cross-package utilities, update imports, and cover the helper with CLI tests.
