+++ src/cli/cli.js
@@ -43,10 +43,8 @@ import {
     toNormalizedLowerCaseString,
     toNormalizedLowerCaseSet
 } from "../shared/string-utils.js";
-import {
-    isPathInside,
-    collectAncestorDirectories
-} from "../shared/path-utils.js";
+import { isPathInside } from "../shared/path-utils.js";
+import { collectAncestorDirectories } from "./lib/ancestor-directories.js";
 
 import {
     CliUsageError,
